chore: Move stuff to Native project
