+++ src/AutoCursorLock.App/App.xaml
@@ -1,9 +1,9 @@
-﻿<Application x:Class="AutoCursorLock.App"
+﻿<Application x:Class="AutoCursorLock.App.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:AutoCursorLock"
              StartupUri="Views/MainWindow.xaml">
     <Application.Resources>
-         
+
     </Application.Resources>
 </Application>
