+++ src/AutoCursorLock.App/Views/AboutWindow.xaml
@@ -1,25 +1,34 @@
-﻿<Window x:Class="AutoCursorLock.AboutWindow"
+﻿<Window x:Class="AutoCursorLock.App.AboutWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:AutoCursorLock"
         mc:Ignorable="d"
-        Title="About" Height="200" Width="300">
-    <Grid x:Name="mainGrid" Margin="5">
+        Title="About"
+        Height="200"
+        Width="300">
+    <Grid x:Name="mainGrid"
+            Margin="5">
         <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
 
-        <Image Grid.Row="0" Source="favicon.ico" Height="64" Width="64" HorizontalAlignment="Left"/>
-        <TextBlock Grid.Row="1" Text="AutoCursorLock" />
+        <Image Grid.Row="0"
+                Source="favicon.ico"
+                Height="64"
+                Width="64"
+                HorizontalAlignment="Left"/>
+        <TextBlock Grid.Row="1"
+                Text="AutoCursorLock"/>
         <WrapPanel Grid.Row="2">
-            <TextBlock Text="v" />
-            <TextBlock Text="{Binding Version, FallbackValue=X.X.X}" />
+            <TextBlock Text="v"/>
+            <TextBlock Text="{Binding Version, FallbackValue=X.X.X}"/>
         </WrapPanel>
-        <TextBlock Grid.Row="3"><Hyperlink NavigateUri="https://github.com/James-LG/AutoCursorLock" RequestNavigate="Hyperlink_RequestNavigate">https://github.com/James-LG/AutoCursorLock</Hyperlink></TextBlock>
+        <TextBlock Grid.Row="3"><Hyperlink NavigateUri="https://github.com/James-LG/AutoCursorLock"
+                    RequestNavigate="Hyperlink_RequestNavigate">https://github.com/James-LG/AutoCursorLock</Hyperlink></TextBlock>
     </Grid>
 </Window>
