+++ src/AutoCursorLock.App/Views/AboutWindow.xaml.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) James La Novara-Gsell. All Rights Reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace AutoCursorLock
+namespace AutoCursorLock.App
 {
     using System.Diagnostics;
     using System.Windows;
