+++ src/AutoCursorLock.App/Views/SelectKeyDialog.xaml
@@ -1,20 +1,28 @@
-﻿<Window x:Class="AutoCursorLock.SelectKeyDialog"
+﻿<Window x:Class="AutoCursorLock.App.SelectKeyDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:AutoCursorLock.Extensions"
         mc:Ignorable="d"
-        Title="SelectKeyDialog" Height="100" Width="100">
+        Title="SelectKeyDialog"
+        Height="100"
+        Width="100">
     <Window.Resources>
-        <local:KeyToStringConverter x:Key="KeyConverter" />
+        <local:KeyToStringConverter x:Key="KeyConverter"/>
     </Window.Resources>
-    <Grid x:Name="mainGrid" PreviewKeyDown="MainGrid_PreviewKeyDown" Focusable="True">
+    <Grid x:Name="mainGrid"
+            PreviewKeyDown="MainGrid_PreviewKeyDown"
+            Focusable="True">
         <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
-        <TextBlock Grid.Row="0" Text="{Binding SelectedKey, Converter={StaticResource KeyConverter}, FallbackValue='Press Any Key'}"  HorizontalAlignment="Center" VerticalAlignment="Center"/>
-        <Button Grid.Row="1" Click="Button_Click">Ok</Button>
+        <TextBlock Grid.Row="0"
+                Text="{Binding SelectedKey, Converter={StaticResource KeyConverter}, FallbackValue='Press Any Key'}"
+                HorizontalAlignment="Center"
+                VerticalAlignment="Center"/>
+        <Button Grid.Row="1"
+                Click="Button_Click">Ok</Button>
     </Grid>
 </Window>
