+++ src/AutoCursorLock.App/Views/SelectKeyDialog.xaml.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) James La Novara-Gsell. All Rights Reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace AutoCursorLock
+namespace AutoCursorLock.App
 {
     using System.ComponentModel;
     using System.Windows;
