--- src/AutoCursorLock.Native/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace AutoCursorLock.Native;
-
-public class Class1
-{
-
-}
