+++ src/AutoCursorLock.Native/ModifierKey.cs
@@ -1,0 +1,35 @@
+﻿// Copyright (c) James La Novara-Gsell. All Rights Reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace AutoCursorLock.Native;
+
+/// <summary>
+/// Represents a modifier key for a hot key.
+/// </summary>
+public enum ModifierKey
+{
+    /// <summary>
+    /// No modifier key.
+    /// </summary>
+    None,
+
+    /// <summary>
+    /// The "Alt" key.
+    /// </summary>
+    Alt,
+
+    /// <summary>
+    /// The "Ctrl" key.
+    /// </summary>
+    Control,
+
+    /// <summary>
+    /// The "Shift" key.
+    /// </summary>
+    Shift,
+
+    /// <summary>
+    /// The "Windows" key.
+    /// </summary>
+    Windows,
+}
